+++ bioblend/galaxy/objects/__init__.py
@@ -1,2 +1,3 @@
 from .galaxy_instance import GalaxyInstance  # noqa: F401
-from .wrappers import *  # noqa: F401,F403
+
+__all__ = ("GalaxyInstance",)